requirements.txt
certifi>=2018.4.16
chardet>=3.0.4
diff-match-patch>=20121119
dj-database-url>=0.4.2
Django>=2.0.5
django-admin-rangefilter>=0.3.6
django-crispy-forms>=1.7.0
django-datetime-widget>=0.9.3
django-import-export>=0.7.0
django-pandas>=0.5.1
django-related-admin>=0.5.0
djangorestframework>=3.8.2
et-xmlfile>=1.0.1
gunicorn>=19.7.1
idna>=2.6
jdcal>=1.3
Jinja2>=2.9.6
MarkupSafe>=1.0
mysqlclient>=1.3.12
numpy>=1.14.4
odfpy>=1.3.6
olefile>=0.44
openpyxl>=2.4.9
pandas>=0.23.0
Pillow>=4.2.1
psycopg2>=2.7.3.1
python-dateutil>=2.7.3
pytz>=2017.3
PyYAML>=3.12
requests>=2.18.4
six>=1.11.0
tablib>=0.12.1
unicodecsv>=0.14.1
urllib3>=1.22
xlrd>=1.1.0
xlwt>=1.3.0
